a bunch of refactoring chaning DatalogChannels / DatalogChannelTypes to be referenced by channelName instead of arrayIndex, to normalize reference of elements throughout app. So much refactoring I almost forgot why :) Part of this was to deal with unknown channel names and also to allow datalogs to have different channels names from other datalogs and not have them collide

+++ src/controls/treelistctrl.cpp
@@ -3123,7 +3123,7 @@ wxTreeItemId wxTreeListMainWindow::GetPrevSibling (const wxTreeItemId& item) con
     // get index
     wxArrayTreeListItems& siblings = parent->GetChildren();
     size_t index = siblings.Index(i);
-    wxASSERT (index != wxNOT_FOUND); // I'm not a child of my parent?
+    wxASSERT ((int)index != wxNOT_FOUND); // I'm not a child of my parent?
     return (index >= 1)? wxTreeItemId(siblings[index-1]): wxTreeItemId();
 }
